Simplified library by unifying llama_embed.cpp file
