+++ .github/codeql-config.yml
@@ -1,0 +1,4 @@
+paths-ignore:
+  - llama.cpp
+  - whisper.cpp
+  - stable-diffusion.cpp
