+++ .github/workflows/codeql.yml
@@ -54,6 +54,7 @@ jobs:
         with:
           languages: ${{ matrix.language }}
           build-mode: ${{ matrix.build-mode }}
+          config-file: .github/codeql-config.yml
 
       - name: Build C/C++ sources
         if: matrix.language == 'c-cpp'
